support multi-card, refactor code
